Save width and height of original image in the image model
This info is important when handling multiple images which don't
have the same aspect ratio and aren't square, eg with 300x200 and
300x250. If you want the thumbnails below the main image, you need
to allocate a box based on the max width and height of the images
in the set, so here you need to set the height to at least 250.

The w/h for the original image are used (paperclip produces this
format by default), though typically you will want to compute the
aspect ratios and apply them to the various other sizes of image.