Additional fix for 660 - we must ignore the address ids from the form
The controller must control the id of the record to write to, since the
form only passes back the id of the ship/bill address that existed when
the form was first loaded, and this is unchanged despite any ajax calls.

So, if the address was new, or if we've switched from a user's default
to a new record, the id will NOT be correct. It must be over-written
with the id of the current active address (ie order.{bill,ship}_address.

Remark: same situation applies when we use nested attributes and ajax,
where the ajax calls can update or replace objects on the form.