217 - add initializer for Paperclip options, and set log default to false
The initializer contains details of the main settings for paperclip,
and specifically sets the :log setting to false (its default in
paperclip is true). Having these settings in an initializer provides an
easy place for developers to change them.

The :whiny setting has been moved out of the image model.

NB there's a few messages which paperclip logs without checking the
:log setting, but these are mainly for adding/deleting images.

[#217 state:resolved milestone:0.8.99]