480 and 504 - improve handling of lookup failures
The products, orders, taxons controllers now all catch invalid ids and
call a handler. If it is defined for the relevant controller, the
handler will call the method "object_missing(id)", else it will call
the generic "render_404(exception)" method.

"object_missing(id)" can be defined in a site to provide customized
recovery from an invalid or out of date URL, eg to redirect to the
home page with a flash message, or to provide a controller-specific
404 page.

This behaviour can be used with any controller that expected to load
an object based on params[:id]. Just include the following line in
the controller:

    prepend_before_filter :reject_unknown_object

[#504 state:resolved milestone:0.8.99]
[#480 state:resolved milestone:0.8.99]