673 - fix broken basket updates
This was due to a reload in update_totals wiping any changes from
form updates etc.

I've opted for the simplest fix for now - remove the reload
AND fixed the need for the reload. update_line_items was
deleting records but leaving the line_items list unchanged.
It now removes the defunct records from the list, thus removing
the need for a reload.

Dropping the reload should only cause problems if line items
get changed independently of the parent order, and I don't
believe this occurs in the code.

However: rethinking the upload vs save situation is needed, to
make this more robust. (Eg can we rely on the order of filters?)