557 - reduce silent failures in checkout process
This fix gets the save filters to return false if a
problem is detected (and so halts the filter chain
and save operation), and adds extra checking in the
controllers to report if problems are detected.

It is expected that the "unexpected error" messages should only
arise from system bugs, and thus only rarely be seen by users.

Additionally, problems in complete_order are caught and
written to the log. Confirmation emails are now sent before
the inventory change and final save, just in case there is
a bug-related failure later on.

[#557 state:resolved]