1204 - relax the ordering-scope predicate for PGs
To support manipulation of unsaved PGs, I've shifted the
test for an ordering scope back to Ruby (was a named scope).

Also relaxed the test too - it now looks for ascend_by/descend_by
OR having :order set in the :find scope. This allows scopes to
be used which aren't in the restricted list for the admin interface.

The Taxon scopes have also been tweaked: neither scope now gives an :order
since this often interferes with the chosen ordering

[#1204 state:resolved]