Explicity define the @orders collection on which to paginate in the admin/orders/index view
This is helpful for extension creators who wish to render this template
from within a controller other than the Admin::OrdersController. By
default, the will_paginate method derives the collection name from the
controller name which would mistakenly be 'orders' in an extension that
renders this template. By explicitly setting the collection to
@orders the template will always render correctly, regardless of the controller that renders it.

This is a trivial change that I don't believe needs testing as @collection
is already tested well in resource_controller.

[#1235 state:resolved]