Fixes changing quantity in cart and update causing item to be removed from the cart
That was actually a side effect of my replacement of link_to by button_for on the "Empty cart" button.
While that is, imo, a better option, in this case, because it was already inside of a form, it didn't work correctly.

button_for wrapps the button into a form, the nesting of the "empty cart" form inside of the cart form caused the browser to think it was a single form, thus, sending the _method=delete key on the submit (update) request.

I turned it back to link_to and added the new Rails3 UJS (jQuery driver) into Spree, which is something we would have to do sooner or later :)