[core] Fix 1.9 bug: Join array into proper string in ProductScope#apply_on
When you call self.arguments it returns an array like [shirt]. In Ruby 1.8.7 when this object is converted to a string it is printed as shirt.

The api/features/products.feature's scenarios all pass in 1.8.7 because of this.

But in 1.9.2 it's actually written as ["shirt"]. A subtle difference, but it causes problems where it attempts to find a product where the name contains that 'inspected' version of the arguments, rather than the arguments literally.

By calling .join(""), we work around this problem.

That was fun to debug