[core] fix failing order spec
On a given day of the week, when the sun is just at the right angle, the outstanding_balance method would return a number like -2.0000000000001, which, of course, is not -2.00 like the test says.

What we care about, as people who deal with money, is that this figure is at least within 1/10th of 1 cent of the correct amount. Anything after that becomes irrelevant.