Base controller decorator MUST be loaded before admin_controller_decorator
This is so the rescue_from for CanCan::AccessDenied is loaded into Spree::BaseController first.

Otherwise, what happens is that Spree::Admin::BaseController gets defined, which loads Spree::BaseController and inherits its class_attributes from that. One of these is rescue_handlers, which at this stage is an empty array. This means that when this controller (or, more likely, one of its descendants) raises the CanCan::AccessDenied exception, the rescue handler is not registered and so the exception causes a *real* exception.

By loading it before hand, the rescue_from call runs on Spree::BaseController before Spree::Admin::BaseController is loaded, meaning that when Spree::Admin::BaseController is *finally* loaded it inherits the rescue_handlers attribute from Spree::BaseController, so that when the CanCan::AccessDenied exception is thrown it is correctly rescued from.

Thanks to adhlssu07 for pointing out this issue