Preferences don't need to be loaded as before_initialize, load them just before spree.environment
Because this causes an issue (#881) where it attempts to load a class that is unloadable at that point in time. What you're attempting to do here if you use before_initialize is that you're attempting to set something up before Rails has really had a chance to do much at all. This is *far too early* in the process to be doing these things.

I think the first point where we need to do this is right before we set up spree.environment.

Ping @cmar.