Move advertised back to being a class method, rather than a scope. Fixes #1071.
If it's a scope, when this model is loaded it attempts to find the spree_activators table (as the scope is evaluated then and there). There is a possibility (see #1071) where this table may not exist at the point of the model being loaded.

By making it a class method, we are evaluating the scope only when the method is called, rather than when the class is loaded.