Correct sorting parameter in admin/products_controller
This was causing PostgreSQL builds to fail intermittently because when PostgreSQL returns data that has no ORDER clause, it will do so in an indeterminite order... the 'I'll do whatever I want to because I'm adult' data return algorithm, I think. By defaulting this to sort by name *properly* it'll always return it name ASC first and therefore won't make the test blow up sometimes and not others.