Move admin translations out of assets and into a special view
That special view is app/views/spree/admin/shared/_translations.html.erb. This file is loaded along with the assets in the <head> tag of the admin layout and defines some JavaScript variables under the 'Spree.translations' 'namespace' which can then be used inside the admin assets.

This *may* cause #1906 to no longer be a problem.