Adds tracking_url field to shipping methods
This feature adds the ability for tracking numbers to be clickable throughout
the system.

Without it, the customer needs to locate the tracking website themselves and
copy and paste the tracking number (except gmail users). It also effects admins
who are trying to investigate the history of a shipment, it would be helpful if
they could just click on the tracking and open up the tracking site.

Fixes #2644