Fix for the edit and create menu items in the taxonomy tree
When clicking edit in the context menu of the taxonomy tree, the browser is redirected to
the edit view of the taxon. Browser should not follow the api route, but the admin route.

Also fix an error in handle create: upon success a taxon object is returned, so id should be called
on the returned object.