Greatly reduce number of drawn routes on Spree::Core::Engine
By prepending routes as we were doing before, Rails was not checking to see if the routes already existed and therefore drew over 1,000 routes for Spree and its things. This, of course, would take time and RAM.

By calling draw, Rails checks and therefore doesn't draw the routes if they're already there.

Thanks to @parndt for pointing this out.