Fixing the problem with translate with relative path.
This commit fix #3176 and add #translate method to Spree module.

To fix this issue, was necessary adding a monkey patch to initialize
the i18n after load the spree base controller.

Also, I changed the method #t. Now this method is an alias to #translate.

In my concept, we need to follow the same convention of i18n and
translation helpers of rails.

So, the developers can use #t and #translate without any problems.

Fixes #3257
Fixes #3176