Delete all inventory units (rather than destroy)
When a shipment was previously destroyed, it would call destroy on each InventoryUnit object. This took a long time if there were a lot of inventory units. There's no need to call destroy, as there's no callbacks in this model for destroy. All shipment needs to do is *DELETE ALL* its inventory units, and so this is what's going to happen now.