Dont update shipments and payment in order cart
Adding a product to cart on a fresh spree install currently runs 41 SQL
queries. This change makes the POST orders/populate request run 34 SQL
queries.

So each time the store needs to run `order.update!` and the order is not
completed yet we run 7 less queries. That could make a big difference
when lots of promotions are involved.

Fixes #3513

Conflicts:
	core/app/models/spree/order_updater.rb