Do-over "Add Stock Movement" form
If there are a large number of stock items for this stock location, the page can take some seconds to load, as mentioned in this thread:

https://groups.google.com/forum/#!topic/spree-user/nu0_NlNGPV8

To make that stop being a problem, I've switched that UI to using select2 to load stock items 50 at a time. I've also imporved the UI to show more information about the product, such as its SKU and its options text.